requirements.txt
Flask==2.0.2
Flask-WTF==0.15.1
mysql-connector-python==8.0.28
